requirements.txt
azure-cognitiveservices-nspkg==3.0.1
azure-cognitiveservices-vision-face==0.4.0
azure-cognitiveservices-vision-nspkg==3.0.1
azure-common==1.1.23
azure-nspkg==3.0.2
certifi==2018.10.15
chardet==3.0.4
cognitive-face==1.5.0
enum34==1.1.6
idna==2.7
imutils==0.5.1
isodate==0.6.0
msrest==0.6.10
numpy==1.15.2
oauthlib==3.1.0
opencv-python==3.4.3.18
python-dotenv==0.10.3
requests==2.20.0
requests-oauthlib==1.2.0
six==1.12.0
typing==3.7.4.1
urllib3==1.24
